requirements.txt
numpy==2.4.3
matplotlib==3.10.8
pygame-ce==2.5.7
torch==2.11.0
torchvision==0.26.0
imageio==2.37.3